oferta de obras

+++ app/Http/Controllers/Obrasyfinan/Ofertas/Ofe_itemController.php
@@ -34,26 +34,24 @@ class Ofe_itemController extends Controller
         return view('Obrasyfinan.Ofertas.ofeitems.editar',compact('unItem','laObra'));
     }
     public function itemsxoferta ($idobra){
+        
         $laObra = Ofe_obra::find(decrypt($idobra));
         $itemsxobra =  $laObra->getItems ;
         return view('Obrasyfinan.Ofertas.ofeitems.index',compact('laObra','itemsxobra'));
     } 
 
-   
-
-
     public function store(Request $request)
     {
         $validatedData = $request->validate([
-            //'id_obra' => 'required|unique|integer',
             'nom_item' => 'required|min:5|max:80|string',
-            'costo' => 'required|min:1|integer',
-            'por_inc' => 'required|min:1|integer',            
+            'orden' => 'required|min:1|integer',
+            'costo' => 'required|min:0|numeric',
+            'por_inc' => 'required|min:0|numeric',                
         ]);
         $input = $request->all(); 
         $itemOferta = new Ofe_item;
         $itemOferta->idobra = $request->input('idobra');
-        $itemOferta->iditem = $request->input('iditem');        
+        $itemOferta->iditem = $request->input('iditem');
         $itemOferta->nom_item = strtoupper($request->input('nom_item'));
         $itemOferta->costo = $request->input('costo');
         $itemOferta->por_inc = $request->input('por_inc');
@@ -62,16 +60,17 @@ class Ofe_itemController extends Controller
         $itemOferta->cod_tipo = 1;        
         $itemOferta->save();
         
-        return redirect()->route('ofeobraitems.itemsoferta',$itemOferta->idobra );
+        return redirect()->route('ofeobraitems.itemsoferta',encrypt($itemOferta->idobra) );
     }
 
     public function update(Request $request, $idItem)
     {
         $validatedData = $request->validate([
             //'id_obra' => 'required|unique|integer',
+            
             'nom_item' => 'required|min:5|max:80|string',
-            'costo' => 'required|min:1|integer',
-            'por_inc' => 'required|min:0|integer',            
+            'costo' => 'required|min:0|numeric',
+            'por_inc' => 'required|min:0|numeric',            
         ]);
         $input = $request->all(); 
         $unItem = Ofe_item::find($idItem);
@@ -83,15 +82,24 @@ class Ofe_itemController extends Controller
         $unItem->por_inc = $request->input('por_inc');
         $unItem->orden = $request->input('orden');
         $unItem->codigo = 1;
-        $unItem->cod_tipo = 1; 
+        $unItem->cod_tipo = $request->input('cod_tipo');; 
         $unItem->save();
-        return redirect()->route('ofeobraitems.itemsoferta',$idobra);
+        return redirect()->route('ofeobraitems.itemsoferta',encrypt($idobra));
     }
     public function destroy($itemId )
-    {        
-        $unItem = Ofe_item::find($itemId)->first();
-        $idobra =$unItem->getObra->idobra;        
-        $unItem->delete();        
-        return redirect()->route('ofeobraitems.itemsoferta', $idobra );
+    {   
+        $unItem = Ofe_item::find($itemId);
+        $subitems = $unItem->getSubItems;
+        $cronograma = $unItem->getCronograma;
+
+        $idobra =$unItem->getObra->idobra;
+        
+        if(  !empty($subitems) || !empty($cronograma)){
+            return redirect()->route('ofeobraitems.itemsoferta', encrypt($idobra)  )->with('error','El item se encuentra en Cronogramas o posee Subitems. Debe eliminarlos para continuar');
+        }else{
+            $unItem->delete(); 
+            return redirect()->route('ofeobraitems.itemsoferta', encrypt($idobra) )->with('mensaje','Item  borrado con éxito!.');
+        }
+        
     }
 }
